Fix MVar blocking
If we try to atomically clear the queue,
we can sometimes never finish clearing it.
This will cause the `finalize` mvar to be blocked for too long.
The runtime system things the mvar is blocked indefinitely,
and kills everything.

Rather, we should not try to atomically clear the queue.
Let each worker finish when they finish.
And just wait until everything is done.