Split out test targets
It looks like,
when we switch jobs in CI the new checkout has newer timestamps
than what's in the make cache.
Rather than dealing with that anymore, we alter the targets.

We can change `test` to ensure builds are up to date.
Then we can leave the individual test targets to run whatever is there.

In CI this means we can set the workflow to run the tests separately.