Setup `nix`
We start with a very basic setup, laying the groundwork for the future.
When we start needing tools to be available,
we can add them to the `buildInputs`.
Hopefully, this will mean that we can keep the environment stable,
reproducible, and tracked in the history.

We also use `direnv` to make loading the `nix` environment easier.
This should cache all of the environment variables from nix,
and load them when we enter the directory.
It's not strictly a requirement, but makes things much easier.