Move `rollbar-hs` from make to shake
We're starting down the path of building everything with one build tool.
Hopefully it doesn't continue growing at nearly three times the size.
Though, we are doing quite a few more checks than the makefile was.

My gut says that once we get a few more underneath it,
it can start to make sense to define generalized rules for building.
Then, each package can configure the build with their few differences.
Finally, read in all the configs on the fly and dynamically create rules.

That's the plan at least...