Move `wai-middleware-rollbar` to shake
We didn't have any sort of build anything before.
Slotting into the setup we've got now is already paying off.
The check noticed a long-standing issue with the cabal file!

We're not really getting the benefit of using a single repo between
the packages yet.
In specific, `wai-middleware-rollbar` depends on `rollbar-hs`.
This dependency is tracked through cabal, rather than through shake.

If we change the local `rollbar-hs` package,
we won't know about how it affects `wai-middleware-rollbar`
until after we update the cabal file and rebuild everything.

What would be better is if a change to the local `rollbar-hs` package
would directly affect `wai-middleware-rollbar`.
Within the world of cabal files, it's unclear how to express that.
We might have to ditch them for development, use shake,
and Greenspun our own manifests.
Not an ideal situation.
Maybe the new cabal stuff can help here?