Make the `builddir` directory explicit
We were relying on the directory being the default.
Not a bad thing when we don't care where artifacts end up
(for example, if we're not writing a Shakefile.hs).
Since we care greatly where these artifacts are
(otherwise rules can be inconsistent),
we should make that explicitly.

This also makes it easier to change the `builddir` later on.
For example,
we have to remember to keep updating the `clean` rule for package.
Rather than that, we can put all the `builddir`s in `_build`
and never have to update again for these Haskell packages.