Move all the dist directories to `_build`
We can manage this easier as more packages are added if all of the
build artifacts are in one location.
This seems to work out pretty alright for cabal-based builds.
It doesn't look like it cares where the build directory is.
Probably other language and build tools won't be so open to the idea.