Move configuration out to files
We're using Dhall for this because it allows communicating types.

This configuration is pretty ad-hoc,
and I don't want to have to try and write prose to explain it,
nor do I want to try and remember what it is supposed to be.

Rather offload this to Dhall, so it can keep things in check.

In any case, the configuration is outside of the build file.
Hopefully this makes pulling in another package easier.
Should be enough to read-tree it into the repo,
dump a `shake.dhall` into the directory,
and be good to go.

Probably won't work like that, because there were only three examples.
So, I guess good luck future me!