Determine rebuilds more intelligently
Relying on timestamps works well if the file system never changes.
In CI, the file system changes on each build.
If we cache the `_build` directory and use it in subsequent builds,
it will have timestamps older than those subsequent builds.

Shake provides alternate ways of comparing files.
One way is to check the timestamp first and digest second on "input"s,
and only the timestamp for "output"s.

"Input"s are files that shake does not build.
"Output"s are files that shake does build.

Let's hope it works as laid out above.