Split the caches
The nix cache takes much longer to warm.
While it doesn't really make a difference for overall time,
knowing what part is taking however long can be beneficial.

Another benefit to splitting the cache is that we can invalidate
part of the cache without affecting all of it.
If the build cache goes sour,
we can kill it without require a full rebuild of nix.