Cache the hash of the build artifacts
The idea here is that `Shakefile.hs` doesn't (and shouldn't)
change enough to be useful as a cache key.
Ideally, `Shakefile.hs` should reach a stable state.

Instead, we want something to cache that changes when `_build` does.
Naively, `_build` is that thing.
We read the files in `_build`, hash them, and track the cache.
To ensure we use the latest cache,
we check the hashed cache file is updated in CI.

Hopefully, this won't be a whole bunch of churn.
Giving how haphazardly it was thrown together, it probably will be.