Track most of the files that matter
We cannot rely on the build artifacts being the same across machines.
This is maybe more appalling than anything else.
But, actually not that unexpected.

We can try to cache something else, instead.
We can cache the files that create the build artifacts.
We're hoping that if build artifacts aren't consistent across machines,
at least they're consistent between builds.

We haven't changed the cache yet, but will shortly.