Format all Haskell files the same way
At least, we're formatting all of the input files the same way.
We might be slipping on which files are formatted properly.
But, the move should be to track those files in the config.
Precision is key here.

Also, we fix a problem with the test inputs.
We were missing all of them because we added an extra directory:
the suite name.
Should try to figure some way around missing entire directories.