Re-add `Shakefile.hs` as part of the cache key
The `Shakefile.hs` file is part of the hash in `.circleci/config`.
However,
we can't make the distinction between it changing and the input files.
Even if all of the inputs change,
we still want to hit a good cache if `Shakefile.hs` hasn't changed.

Let's say there are two branches going,
and both are uploading caches to
`v1-build-{{checksum ".circleci/config" }}`.
If one build misses the checksum part (because inputs changed),
it will think the `v1-build-` matches (with the changed inputs).

The unfortunate part is,
it's possible the `Shakefile.hs` file is also changing on both branches.
If that's the case, the cache isn't worth much.
Within the `Shakefile.hs` file,
we tell shake to rebuild all rules if the `Shakefile.hs` hash changes.
So, we'd think we'd need to rebuild everything even if we have a cache.

We can be a bit more intelligent about caching,
if we add the `Shakefile.hs` file as a fallback in the key.
Now, even if multiple branches are building at similar times,
a cache hit with the `Shakefile.hs` key will have a safe database.