Split the `Shakefile.hs` file by responsibilities
Thee are two reasons for this change:
1. The `Shakefile.hs` was getting too large for me to manage.
2. Every change to `Shakefile.hs` required a full rebuild.

The first reason is subjective, but that's what it is.

The second reason is a bigger pain than it sounds.
Each push to CI that touched `Shakefile.hs` required a full rebuild.
That meant each of those builds was like 4-6 minutes.
That's way too long for a little code as we have here.

More importantly, it's not going to work well when we add more languages.
If we get to five or six Haskell packages,
then decide to add another language,
any of the caching we had would be lost once we touch `Shakefile.hs`.

How does this change get around the full rebuild?
We were using `shakeVersion` to force a rebuild whenever
the `Shakefile.hs` changed.
That's the right way to deal with changes when everything is in one file.
If a rule for a real artifact changes,
it should be rebuilt so the build stays consistent across changes.

What we do now is keep the rules for real artifacts in their own files.
Within the rules, we `need` the file themselves.
If a rule changes, the file has changed, so the rule is re-run.
Since the rules are split by files, the scope of changes is much smaller.

There's still some coarseness to the changes.
If one rule changes, it affects all rules in that file.
But it means that we can add another set of rules,
and it not force the entire repo to rebuild.

When it comes to `phony` rules, they don't need to `need` their file.
Since they're always run all the time,
it doesn't matter if the file changes or not.

The `Shakefile.hs` file can now drop `shakeVersion` because
it carries no useful state in the file anymore.