Switch over to ReaderT
Knowing which `FilePath` is which is confusing, at best.
But also, threading the arguments deep into rules is hard.

As the build grows, it would only get more complex.
ReaderT gives us a good balance, and is no more complex than before.
We have to do a bunch of `lift`ing, but it's not a big deal.