Bring the executables under shake's control
In a similar way to how we want to control libraries and tests,
we also want to control executables.

We might consider making a passing test suite be a `need`.
In that sense, you cannot make an executable with broken behavior.
Seems like a safe bet, for the most part.