Add a rule to put binaries in an easy location
The binaries are currently in the `_build` directory.
That works, but it's not the easiest to locate later on.
To make things easier to find,
we copy them to a new `bin` directory.

The idea here is that the directory hierarchy can be flattened
since we expect each binary to be uniquely named within the repo.