Start working on a Dhall interpreter in go
There are two ideas behind this interpreter:
1. Provide a go interface to Dhall.
2. Make it worlds easier to get a `dhall` binary on any platform.

The first is a nice-to-have.
Lots of YAML these days is backed by go tools.
Having a direct line between the tool and the configuration
should make some ideas easier to deal with.
It might even spur new ideas as well.

The second is a pretty big want for working on different platforms.
The current state of the world is to pull down a pre-compiled binary
if you're on Linux, use brew on OSX/macOS, or try your hand at nix.
In theory, that's all fine, and it can work.
In practice, different versions of operating systems wreak havoc.

For instance, if we're using nix to build `dhall`,
we currently cannot build Dhall newer than 1.16.1 on macOS Mojave.
We can build the newest Dhall on macOS High Sierra, though.

Hopefully, we can use go's wonderful cross-compilation support
to have a binary for every platform very easily.