Add a suite for `BoolEqual`
This was a pretty big change, and a really good suite.
It fleshed out that there were assumptions we'd made about how go works.
Those assumptions were wrong.
Some assumptions that were wrong:
* Equality of pointers will dereference the address.
* Equality of interfaces will work if the interface has a function.
* Equality of `[]interface{}` will work
* Upcasting interfaces will work

We also had wrong assumptions about `testing.T.Error`/`testing.T.Errorf`.
It doesn't stop the test.
We need `testing.T.Fatal`/`testing.T.Fatalf`, respectively.

We also wrote a straight up bug in `BoolEqual.infer` method.
We expected each side to infer to a `Bool` instead of a `BoolType`.
Plus, we weren't checking that correctly.

Overall, this was a great addition to the set of tests.