Start parsing Dhall source code
This is a huge commit, unfortunately.
We started at the top of the grammar,
and added only what was necessary to get the few `Expression`s we have
to be able to parse.
Good news is that a fairly large chunk of parsing is done.

We should focus on creating a CLI now.
If for no other reason than that it's incredibly verbose to parse.
Having a switch to kill that should be useful.