Throw together a CLI
This thing is a real program now!
We just take the three things we were doing before,
and turn them into formal commands.

Since there's no requirements that a Dhall CLI behaves in a certain way,
we can pretty much do whatever we want with this thing.
Hopefully we can get to feature parity with the `dhall-haskell` version.