Support output as different formats
We're diverging from the `dhall-haskell` CLI here.
With the `dhall-haskell` version, each executable is different.
Here, we're saying that most commands can output in different formats.

Hopefully, this makes it easier on consumers.
There should be just one binary to install and one version to deal with.

We've only got YAML, but we can expand it.

Maybe the `encode` command is actually another output format?