Support some programming languages as output
This is a pretty good smattering of languages that I personally use.
Hopefully, this gives insight that Dhall can compile to most
mainstream programming languages.

It's also pretty interesting that almost none of the `Constant`s
can be rendered to a mainstream language.
Haskell and PureScript being the exceptions, and they can only do `Type`.