Improve parsing performance immensely
We wrote everything as a function, initially.
Semantically, that's fine.
In practice, it meant we were always creating new values.
Each time we would, it would get garbage collected.

This bogged down the parser a huge amount.
For instance, the test suite took about 10 seconds to run.
That's about a second per call to `dhall.Parse`.

Now, we move everything to a variable that can be.
This means we only create the vast majority of things once.

The test suite now runs in about 0.2 seconds.
Not bad for a little bit of rearranging.