Convert some tests to properties
Many of the tests were hardcoded,
which meant we'd not get a good spread of values under test.

The `gopter` package doesn't seem to work well with recursive data,
so it's unclear if this is going to be useful in the long run.
Hopefully, it mitigates the boilerplate we have to write.