Start working on a build system
The idea isn't to replace `pulp` or `spago` or something.
It's to replace `make`.

`shake` is an amazing build system.
But it comes with some baggage,
namely that you need a Haskell compiler in order to run it.
That's fine if every place you intend to use a build system
written in `shake` has Haskell installed,
But GHC (let alone Haskell) is not ubiquitous enough for that.

To get around this limitation, we create a system on top of `shake`.
We can compile this system to a binary and let others use it.
This idea was suggested by the author of `shake`:
https://github.com/ndmitchell/shake/issues/655#issuecomment-468274574.

Currently, it only "works" for PureScript programs.
The way it works is by reading the imports from the `Main` module,
and downloading whatever it needs.
It's not the most parallel way to do this,
but it's probably as good as we'll get given the way PS has shaken out.

This repo is probably going to be the testbed for this build system.
Since this is a monorepo (with multiple languages soon),
lots of edge cases come up often.

Hopefully this works out,
as trying to keep all these languages working together is getting complex.