Start splitting out build and PS stuff
The `Main` module was getting a bit too much to understand.
Since we're trying to address multiple languages and ideas,
it seems like a fair move to start splitting out by each aspect.

The modules aren't right, since `Build.PureScript` depends on `Build`.
We want it to be the other way around,
but we'll have to break some stuff in order to do that.
Also, there's PS specific stuff in `Build`.
That's not right.