Support different targets
Now that we have a binary that can browserify without requiring node,
we can download it and use it to browserify without requiring node!

We use it by allowing two different targets for PureScript programs.
The `Browser` target will pull down the binary and browserify it.
No muss, no fuss.

Endgame here is that these rules move out of `build` and into userland.
It shouldn't be too hard, but it will take some time to get there.