Improve haskellPackages overlay
The `self` argument is the fixed point,
the `super` argument the previous point.
We should use `self` for packages that might be altered
and `super` for things that should be normal (like functions).

We also move the packages up to the overlay itself.
That should help things be a bit clearer as to what's happening:
we're adding packages and later using them.