Update `shake` to `0.17.7`
There's quite a large number of changes since `0.16.4`.
Notably, the introduction of shared databases.
We'll have to see if we get a speed improvement by using that.
There's also improvements to reporting and the API in general.
This is a big change, so it might take a few commits to use it all.