Start using `shakeShare`
This is the caching system that Shake provides.
We put the cache outside of the repo with the idea for now.
It's unclear if it should live within the repo or outside of it.

In any case, this seems to work like expected:
building, cleaning, then building again means the second build is
nearly instantaneous.

The cache seems to take up quite a bit of space.
We might want to see if we can't use smaller files for things like
formatting and linting once this becomes a problem.
Until then, caching!