Change the cache key
Looks like nothing gets written to `~/.cache/shake`.
Might be a permissions thing, since `~/` is `/root` in this case.

Trying with a local directory.
This should work out since local stuff should always be written to.

We also put some unique identifier on the cache key.
Otherwise, the cache is never updated since the key is always the same.