Fix dependencies for Haskell formatting and linting
We were depending on too much.
If we depend on all Haskell files in the `Shake` directory,
Any one of them changing means that every Haskell file in the repo
has to have the rules run again.
We don't want that to be the case.
At most we need to depend on the Haskell file and the rule for
formatting and/or linting.

Since these rules are within the `Shake/Haskell.hs` file,
we depend on that file alone.
It's not the most granular, but it gets to the level that we care about.