Update nixpkgs to 19.03
A new release is out, and we can update to it.

The update to 19.03 changed some dependencies.
We thread that through to the packages we use.

Also,
all of our rollbar-based packages are marked as broken in 19.03.
It's a little weird that any of these packages were in nixpkgs at all.
We didn't add them anytime, they just appeared in there.
It's also a little weird that nothing was communicated about them
being marked broken.
The nixpkgs workflow is an odd one.

The cabal changes are due to the CLI updating to a new API.
The old stuff is deprecated.
New commands are around that explicitly reference the old stuff.
We should move off of the `v1-` stuff as that's supposed to go away.