Add missing derivation for `rollbar-hs`
This was missed from the previous commit.
Part of our rollbar packages being marked broken in 19.03
meant that we had to override the packages with a non-broken package.

nixpkgs marking our rollbar packages as broken was a blessing in disguise.
It forced our hand on being able to build the rest of our packages
that depend on the local version of `rollbar-hs`.
We should do this with all of our local dependencies instead of using
whatever things exist in nixpkgs.