Fix `nexe` command
We weren't able to run `yarn run nexe` correctly.
It's unclear how we had it working in the past,
but it was broken and started showing that recently.
It seems like something is going sideways since it's not being run
directly in a terminal or something?
The hint in:
https://github.com/ndmitchell/shake/issues/212#issuecomment-75426463
was able to point in the correct direction.
I.e. we needed to add `WithStderr False`.