Do not cache `yarn install`
Attempting to cache the output of `yarn install` is hard.
It installs things with symlinks, `shake` doesn't keep symlinks.
`shake` follows the link to the file and replaces the symlinks.

For the time being, we're going to always run a `yarn install`.
It's fairly cheap work, so not a huge loss.