Parse a module in a `string`
Another thing we'll need is to parse the module name.
We cannot rely on the file name to tell us what module it is,
as PureScript doesn't work that way.

Instead, we have to parse the module to figure it out.
It would be nice if this information was available outside of the
source language, but that's probably never going to happen for PS.