Create a compilable program
Up until now, we've only been able to run the code as tests.
We want to distribute a compiled program to people.
We need to have a `main` that `go` can use to compile into a program.

We read in whatever is sent over STDIN, parse it, and graph it.
We'll want to accept files at some point,
but this gets us at least able to use the program for now.