Change `graph` to accept a `[]psModule`
We want to be able to work with any number of modules.
The idea here is that making a graph of the imports of a single
module is useful, but not all that interesting.
The graph becomes more interesting when we can see it with other
modules that are used together.

Once we can see the whole graph,
we can start to understand the program at a different level.