Parse arguments sent in over the CLI
We want to work with more than just STDIN.
We want people to be able to pass a set of files in as arguments,
and parse each file for its graph.

We'll build up that graph as we parse each file.
Once we're done building the graph, we display it at the end.