Start working on a Mikado graph
Mikado graphs are pretty useful. The problem though is that they're not
always easy to maintain in code.

They're supposed to be things you draw on a whiteboard or paper or
something. But often, you want the graph to be stored in a code base.
There could be many reasons for that: distributed team, version control,
ephemeralness, etc.

Instead of writing in plain Graphviz, we can use Dhall to make things
easier. There's still some rough edges to figure out, but this is a good
enough start.